requirements.txt
chromadb
sentence-transformers
google-genai
gradio
python-dotenv
textwrap3
